refactor: cookie login 여부확인을 도메인 로직으로 변경
