+++ src/test/java/webserver/RequestMapperTest.java
@@ -3,7 +3,6 @@ package webserver;
 import controller.*;
 import model.HttpMethod;
 import model.RequestMappingInfo;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
